Gem_Demo2/FactChecker/factCheckMCP/requirements.txt
mcp[cli]>=1.13.1
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
google-generativeai>=0.7.0
tenacity>=8.2.0
python-dotenv>=1.0.0
